reorganize: library service goes into services/ dir
