+++ src/lib/services/library/addBook.ts
@@ -1,6 +1,6 @@
 import { dialog } from "@tauri-apps/api";
 import type { Library } from "./_types";
-import type { ImportableBookMetadata } from "../../bindings";
+import type { ImportableBookMetadata } from "../../../bindings";
 
 export const promptToAddBook = async (
 	library: Library,
