+++ src/lib/services/library/libraryCommsManager.ts
